fix: move ReadMetricFile to prevent import cycle
