+++ plugins/parsers/csv/parser_test.go
@@ -1,4 +1,4 @@
-package csv
+package csv_test
 
 import (
 	"fmt"
@@ -15,6 +15,8 @@ import (
 	"github.com/influxdata/telegraf/metric"
 	"github.com/influxdata/telegraf/plugins/inputs"
 	"github.com/influxdata/telegraf/plugins/inputs/file"
+	"github.com/influxdata/telegraf/plugins/parsers/csv"
+	parser_testutil "github.com/influxdata/telegraf/plugins/parsers/testutil"
 	"github.com/influxdata/telegraf/testutil"
 )
 
@@ -23,7 +25,7 @@ var DefaultTime = func() time.Time {
 }
 
 func TestBasicCSV(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		ColumnNames: []string{"first", "second", "third"},
 		TagColumns:  []string{"third"},
 		TimeFunc:    DefaultTime,
@@ -36,7 +38,7 @@ func TestBasicCSV(t *testing.T) {
 }
 
 func TestHeaderConcatenationCSV(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    2,
 		MeasurementColumn: "3",
 		TimeFunc:          DefaultTime,
@@ -53,7 +55,7 @@ func TestHeaderConcatenationCSV(t *testing.T) {
 }
 
 func TestHeaderOverride(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -74,7 +76,7 @@ func TestHeaderOverride(t *testing.T) {
 
 	testCSVRows := []string{"line1,line2,line3\r\n", "3.4,70,test_name\r\n"}
 
-	p = &Parser{
+	p = &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -92,7 +94,7 @@ func TestHeaderOverride(t *testing.T) {
 }
 
 func TestTimestamp(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -114,7 +116,7 @@ func TestTimestamp(t *testing.T) {
 }
 
 func TestTimestampYYYYMMDDHHmm(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -135,7 +137,7 @@ func TestTimestampYYYYMMDDHHmm(t *testing.T) {
 	require.Equal(t, metrics[1].Time().UnixNano(), int64(1247328300000000000))
 }
 func TestTimestampError(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -152,7 +154,7 @@ func TestTimestampError(t *testing.T) {
 }
 
 func TestTimestampUnixFormat(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -172,7 +174,7 @@ func TestTimestampUnixFormat(t *testing.T) {
 }
 
 func TestTimestampUnixMSFormat(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -192,7 +194,7 @@ func TestTimestampUnixMSFormat(t *testing.T) {
 }
 
 func TestQuotedCharacter(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -209,7 +211,7 @@ func TestQuotedCharacter(t *testing.T) {
 }
 
 func TestDelimiter(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		Delimiter:         "%",
 		ColumnNames:       []string{"first", "second", "third"},
@@ -227,7 +229,7 @@ func TestDelimiter(t *testing.T) {
 }
 
 func TestValueConversion(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount: 0,
 		Delimiter:      ",",
 		ColumnNames:    []string{"first", "second", "third", "fourth"},
@@ -268,7 +270,7 @@ func TestValueConversion(t *testing.T) {
 }
 
 func TestSkipComment(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount: 0,
 		Comment:        "#",
 		ColumnNames:    []string{"first", "second", "third", "fourth"},
@@ -293,7 +295,7 @@ func TestSkipComment(t *testing.T) {
 }
 
 func TestTrimSpace(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount: 0,
 		TrimSpace:      true,
 		ColumnNames:    []string{"first", "second", "third", "fourth"},
@@ -315,7 +317,7 @@ func TestTrimSpace(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, expectedFields, metrics[0].Fields())
 
-	p = &Parser{
+	p = &csv.Parser{
 		HeaderRowCount: 2,
 		TrimSpace:      true,
 		TimeFunc:       DefaultTime,
@@ -332,7 +334,7 @@ func TestTrimSpace(t *testing.T) {
 }
 
 func TestTrimSpaceDelimitedBySpace(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		Delimiter:      " ",
 		HeaderRowCount: 1,
 		TrimSpace:      true,
@@ -359,7 +361,7 @@ abcdefgh        0       2    false
 }
 
 func TestSkipRows(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		SkipRows:          1,
 		TagColumns:        []string{"line1"},
@@ -385,7 +387,7 @@ hello,80,test_name2`
 	require.Equal(t, expectedFields, metrics[0].Fields())
 	require.Equal(t, expectedTags, metrics[0].Tags())
 
-	p = &Parser{
+	p = &csv.Parser{
 		HeaderRowCount:    1,
 		SkipRows:          1,
 		TagColumns:        []string{"line1"},
@@ -411,7 +413,7 @@ hello,80,test_name2`
 }
 
 func TestSkipColumns(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		SkipColumns: 1,
 		ColumnNames: []string{"line1", "line2"},
 		TimeFunc:    DefaultTime,
@@ -430,7 +432,7 @@ func TestSkipColumns(t *testing.T) {
 }
 
 func TestSkipColumnsWithHeader(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		SkipColumns:    1,
 		HeaderRowCount: 2,
 		TimeFunc:       DefaultTime,
@@ -449,7 +451,7 @@ trash,80,test_name`
 }
 
 func TestMultiHeader(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount: 2,
 		TimeFunc:       DefaultTime,
 	}
@@ -464,7 +466,7 @@ func TestMultiHeader(t *testing.T) {
 
 	testCSVRows := []string{"col,col\r\n", "1,2\r\n", "80,test_name\r\n"}
 
-	p = &Parser{
+	p = &csv.Parser{
 		HeaderRowCount: 2,
 		TimeFunc:       DefaultTime,
 	}
@@ -484,7 +486,7 @@ func TestMultiHeader(t *testing.T) {
 }
 
 func TestParseStream(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:     "csv",
 		HeaderRowCount: 1,
 		TimeFunc:       DefaultTime,
@@ -514,7 +516,7 @@ func TestParseStream(t *testing.T) {
 }
 
 func TestParseLineMultiMetricErrorMessage(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:     "csv",
 		HeaderRowCount: 1,
 		TimeFunc:       DefaultTime,
@@ -550,7 +552,7 @@ func TestParseLineMultiMetricErrorMessage(t *testing.T) {
 }
 
 func TestTimestampUnixFloatPrecision(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:      "csv",
 		ColumnNames:     []string{"time", "value"},
 		TimestampColumn: "time",
@@ -579,7 +581,7 @@ func TestTimestampUnixFloatPrecision(t *testing.T) {
 }
 
 func TestSkipMeasurementColumn(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:      "csv",
 		HeaderRowCount:  1,
 		TimestampColumn: "timestamp",
@@ -611,7 +613,7 @@ func TestSkipMeasurementColumn(t *testing.T) {
 }
 
 func TestSkipTimestampColumn(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:      "csv",
 		HeaderRowCount:  1,
 		TimestampColumn: "timestamp",
@@ -643,7 +645,7 @@ func TestSkipTimestampColumn(t *testing.T) {
 }
 
 func TestTimestampTimezone(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"first", "second", "third"},
 		MeasurementColumn: "third",
@@ -666,7 +668,7 @@ func TestTimestampTimezone(t *testing.T) {
 }
 
 func TestEmptyMeasurementName(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:        "csv",
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"", "b"},
@@ -693,7 +695,7 @@ func TestEmptyMeasurementName(t *testing.T) {
 }
 
 func TestNumericMeasurementName(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:        "csv",
 		HeaderRowCount:    1,
 		ColumnNames:       []string{"a", "b"},
@@ -720,7 +722,7 @@ func TestNumericMeasurementName(t *testing.T) {
 }
 
 func TestStaticMeasurementName(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:     "csv",
 		HeaderRowCount: 1,
 		ColumnNames:    []string{"a", "b"},
@@ -747,7 +749,7 @@ func TestStaticMeasurementName(t *testing.T) {
 }
 
 func TestSkipEmptyStringValue(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:     "csv",
 		HeaderRowCount: 1,
 		ColumnNames:    []string{"a", "b"},
@@ -774,7 +776,7 @@ func TestSkipEmptyStringValue(t *testing.T) {
 }
 
 func TestSkipSpecifiedStringValue(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		MetricName:     "csv",
 		HeaderRowCount: 1,
 		ColumnNames:    []string{"a", "b"},
@@ -801,7 +803,7 @@ func TestSkipSpecifiedStringValue(t *testing.T) {
 }
 
 func TestSkipErrorOnCorruptedCSVLine(t *testing.T) {
-	p := &Parser{
+	p := &csv.Parser{
 		HeaderRowCount:  1,
 		TimestampColumn: "date",
 		TimestampFormat: "02/01/06 03:04:05 PM",
@@ -866,7 +868,7 @@ func TestMultipleConfigs(t *testing.T) {
 			}
 
 			// Process expected metrics and compare with resulting metrics
-			expectedOutputs, err := testutil.ReadMetricFile(fmt.Sprintf("testdata/%s/expected.out", f.Name()))
+			expectedOutputs, err := parser_testutil.ReadMetricFile(fmt.Sprintf("testdata/%s/expected.out", f.Name()))
 			require.NoError(t, err)
 			resultingMetrics := acc.GetTelegrafMetrics()
 			testutil.RequireMetricsEqual(t, expectedOutputs, resultingMetrics, testutil.IgnoreTime())
