+++ plugins/parsers/json_v2/parser.go
@@ -1,7 +1,9 @@
 package json_v2
 
 import (
+	"bufio"
 	"fmt"
+	"os"
 	"strconv"
 	"strings"
 	"time"
@@ -9,6 +11,7 @@ import (
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/internal"
 	"github.com/influxdata/telegraf/metric"
+	"github.com/influxdata/telegraf/plugins/parsers/influx"
 	"github.com/tidwall/gjson"
 )
 
@@ -622,4 +625,34 @@ func (p *Parser) convertType(input gjson.Result, desiredType string, name string
 	}
 
 	return input.Value(), nil
+}
+
+func ReadMetricFile(path string) ([]telegraf.Metric, error) {
+	var metrics []telegraf.Metric
+	expectedFile, err := os.Open(path)
+	if err != nil {
+		return metrics, err
+	}
+	defer expectedFile.Close()
+
+	parser := influx.NewParser(influx.NewMetricHandler())
+	scanner := bufio.NewScanner(expectedFile)
+	for scanner.Scan() {
+		line := scanner.Text()
+		if line != "" {
+			m, err := parser.ParseLine(line)
+			// The timezone needs to be UTC to match the timestamp test results
+			m.SetTime(m.Time().UTC())
+			if err != nil {
+				return nil, fmt.Errorf("unable to parse metric in %q failed: %v", line, err)
+			}
+			metrics = append(metrics, m)
+		}
+	}
+	err = expectedFile.Close()
+	if err != nil {
+		return metrics, err
+	}
+
+	return metrics, nil
 }
